Move api to client/api
