+++ src/dataclay/__init__.py
@@ -2,5 +2,6 @@
 Public dataclay functions exported to use (friendly) "from dataclay import ..."
 """
 from dataclay.dataclay_object import DataClayObject, activemethod
+from dataclay.client import api
 
 StorageObject = DataClayObject
